Refactored investment value form handlers into own middleware
