+++ src/apps/investment-projects/middleware/form.js
@@ -4,7 +4,6 @@ const metadataRepo = require('../../../lib/metadata')
 const { getAdvisers } = require('../../adviser/repos')
 const interactionFormattingService = require('../../interactions/services/formatting')
 const {
-  valueLabels,
   interactionsLabels,
   requirementsLabels,
 } = require('../labels')
@@ -13,19 +12,6 @@ const {
   updateInvestmentInteraction,
 } = require('../../interactions/repos')
 const { updateInvestment } = require('../repos')
-
-function populateValueFormMiddleware (req, res, next) {
-  res.locals.form = get(res, 'locals.form', {})
-  res.locals.form.labels = valueLabels.edit
-  res.locals.form.state = Object.assign({}, res.locals.investmentData, {
-    average_salary: get(res.locals.investmentData, 'average_salary.id'),
-  })
-  res.locals.form.options = {
-    averageSalaryRange: metadataRepo.salaryRangeOptions,
-  }
-
-  next()
-}
 
 function populateRequirementsFormMiddleware (req, res, next) {
   res.locals.form = get(res, 'locals.form', {})
@@ -39,32 +25,6 @@ function populateRequirementsFormMiddleware (req, res, next) {
   }
 
   next()
-}
-
-function investmentValueFormPostMiddleware (req, res, next) {
-  res.locals.projectId = req.params.id
-
-  const formattedBody = Object.assign({}, req.body, {
-    average_salary: {
-      id: req.body.average_salary,
-    },
-    client_cannot_provide_total_investment: req.body.client_cannot_provide_total_investment === 'on',
-    client_cannot_provide_foreign_investment: req.body.client_cannot_provide_foreign_investment === 'on',
-  })
-
-  updateInvestment(req.session.token, res.locals.projectId, formattedBody)
-    .then(() => next())
-    .catch((err) => {
-      if (err.statusCode === 400) {
-        res.locals.form = get(res, 'locals.form', {})
-        res.locals.form.errors = err.error
-        res.locals.form.state = req.body
-
-        next()
-      } else {
-        next(err)
-      }
-    })
 }
 
 async function populateInteractionsFormMiddleware (req, res, next) {
@@ -149,9 +109,7 @@ function investmentRequirementsFormPostMiddleware (req, res, next) {
 }
 
 module.exports = {
-  investmentValueFormPostMiddleware,
   investmentRequirementsFormPostMiddleware,
-  populateValueFormMiddleware,
   interactionDetailsFormPostMiddleware,
   populateInteractionsFormMiddleware,
   populateRequirementsFormMiddleware,
